+++ src/com/aventura/math/perspective/Frustum.java
@@ -35,15 +35,15 @@ import com.aventura.tools.tracing.Tracer;
  */
 public class Frustum extends Perspective {
 	
-	public Frustum(double left, double right, double bottom, double top, double near, double far) {
+	public Frustum(float left, float right, float bottom, float top, float near, float far) {
 		
 		if (Tracer.function) Tracer.traceFunction(this.getClass(), "New Frustrum perspective");
 		if (Tracer.info) Tracer.traceInfo(this.getClass(), "Perspectives range: left: "+left+" right: "+right+" bottom: "+bottom+" top: "+top+" near: "+near+" far: "+far);						
 		
-		double[][] array = { { 2*near/(right-left) , 0.0                  , (right+left)/(right-left), 0.0                   },
-				 			 { 0.0                 , 2*near/(top-bottom)  , (top+bottom)/(top-bottom), 0.0                   },
-				 			 { 0.0                 , 0.0                  , (near+far)/(near-far)    , 2*near*far/(near-far) },
-				 			 { 0.0                 , 0.0                  , -1.0                     , 0.0                   } };
+		float[][] array = { { 2*near/(right-left)  , 0.0f                  , (right+left)/(right-left) , 0.0f                   },
+				 			{ 0.0f                 , 2*near/(top-bottom)   , (top+bottom)/(top-bottom) , 0.0f                   },
+				 			{ 0.0f                 , 0.0f                  , (near+far)/(near-far)     , 2*near*far/(near-far)  },
+				 			{ 0.0f                 , 0.0f                  , -1.0f                     , 0.0f                   } };
 		
 		try {
 			this.setArray(array);
